requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
PyYAML>=6.0
tqdm>=4.65.0
# optional: wandb
